+++ src/main/java/org/morphix/reflection/GenericClass.java
@@ -1,7 +1,20 @@
+/*
+ * Copyright 2025 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
 package org.morphix.reflection;
 
 import java.lang.reflect.ParameterizedType;
 import java.lang.reflect.Type;
+import java.util.Objects;
 
 /**
  * The purpose of this class is to enable capturing and passing a generic type. This is achieved by creating an inline
@@ -31,7 +44,7 @@ public abstract class GenericClass<T> {
 	 */
 	protected GenericClass() {
 		Class<?> genericClassSubclass = Classes.findSubclass(GenericClass.class, getClass());
-		this.type = GenericType.getGenericParameterType(genericClassSubclass, 0);
+		setType(GenericType.getGenericParameterType(genericClassSubclass, 0));
 	}
 
 	/**
@@ -40,7 +53,7 @@ public abstract class GenericClass<T> {
 	 * @param type captured type
 	 */
 	private GenericClass(final Type type) {
-		this.type = type;
+		setType(type);
 	}
 
 	/**
@@ -64,7 +77,7 @@ public abstract class GenericClass<T> {
 	 */
 	public void setType(final Type type) {
 		if (!(type instanceof ParameterizedType)) {
-			throw new ReflectionException("Type must be a " + ParameterizedType.class);
+			throw new ReflectionException("Generic argument type must be a generic class (ParameterizedType), but got: " + type);
 		}
 		this.type = type;
 	}
@@ -104,7 +117,7 @@ public abstract class GenericClass<T> {
 		if (this == obj) {
 			return true;
 		}
-		return obj instanceof GenericClass<?> that && type.equals(that.type);
+		return obj instanceof GenericClass<?> that && Objects.equals(type, that.type);
 	}
 
 	/**
@@ -112,7 +125,7 @@ public abstract class GenericClass<T> {
 	 */
 	@Override
 	public int hashCode() {
-		return type.hashCode();
+		return Objects.hashCode(type);
 	}
 
 	/**