+++ src/main/java/org/morphix/reflection/GenericType.java
@@ -100,7 +100,7 @@ public class GenericType implements ParameterizedType {
 	 */
 	public static GenericType of(final ParameterizedType parameterizedType) {
 		return of(
-				(Class<?>) parameterizedType.getRawType(),
+				JavaObjects.cast(parameterizedType.getRawType()),
 				Objects.requireNonNull(parameterizedType.getActualTypeArguments(), "actual type arguments"),
 				parameterizedType.getOwnerType());
 	}
@@ -134,7 +134,7 @@ public class GenericType implements ParameterizedType {
 	 */
 	public static <T extends Type> T getGenericParameterType(final Class<?> cls, final int index) {
 		if (index < 0) {
-			throw new ReflectionException("index cannot be negative.");
+			throw new ReflectionException("Generic parameter type index cannot be negative, received: " + index);
 		}
 		if (isNotGenericClass(cls)) {
 			throw new ReflectionException(cls.getCanonicalName() + " is not a generic class");
@@ -142,8 +142,8 @@ public class GenericType implements ParameterizedType {
 		ParameterizedType parameterizedType = (ParameterizedType) cls.getGenericSuperclass();
 		Type[] actualTypeArguments = parameterizedType.getActualTypeArguments();
 		if (index >= actualTypeArguments.length) {
-			throw new ReflectionException(cls.getCanonicalName() + " class has only " + actualTypeArguments.length
-					+ " generic arguments. Index: " + index + " is out of bounds.");
+			throw new ReflectionException("Cannot extract generic parameter type at index " + index + " from " + cls.getCanonicalName()
+					+ " because it has only " + actualTypeArguments.length + " generic parameter(s)");
 		}
 		Type genericType = actualTypeArguments[index];
 		return JavaObjects.cast(genericType);
@@ -230,10 +230,10 @@ public class GenericType implements ParameterizedType {
 		if (o == this) {
 			return true;
 		}
-		if (o instanceof GenericType that) {
-			return Objects.equals(that.rawType, this.rawType)
-					&& Arrays.equals(that.arguments, this.arguments)
-					&& Objects.equals(that.ownerType, this.ownerType);
+		if (o instanceof ParameterizedType that) {
+			return Objects.equals(that.getRawType(), this.rawType)
+					&& Arrays.equals(that.getActualTypeArguments(), this.arguments)
+					&& Objects.equals(that.getOwnerType(), this.ownerType);
 		}
 		return false;
 	}
@@ -249,7 +249,7 @@ public class GenericType implements ParameterizedType {
 	}
 
 	/**
-	 * @see #hashCode()
+	 * @see #toString()
 	 */
 	@Override
 	public String toString() {
@@ -260,7 +260,7 @@ public class GenericType implements ParameterizedType {
 			sb.append("$");
 			if (ownerType instanceof ParameterizedType ot) {
 				// Find simple name of the nested type by removing the shared prefix with an owner.
-				sb.append(rawType.getName().replace(((Class<?>) ot.getRawType()).getName() + "$", ""));
+				sb.append(rawType.getName().replace(ot.getRawType().getTypeName() + "$", ""));
 			} else {
 				sb.append(rawType.getSimpleName());
 			}