+++ src/test/java/org/morphix/utils/Tests.java
@@ -14,7 +14,9 @@ package org.morphix.utils;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
+import java.io.FileInputStream;
 import java.lang.reflect.InvocationTargetException;
+import java.util.Properties;
 
 import org.morphix.lang.JavaObjects;
 import org.morphix.reflection.Constructors;
@@ -27,11 +29,21 @@ import org.morphix.reflection.ReflectionException;
  */
 public interface Tests {
 
-	public static <T extends Throwable> T verifyDefaultConstructorThrows(final Class<?> cls) {
+	static <T extends Throwable> T verifyDefaultConstructorThrows(final Class<?> cls) {
 		ReflectionException reflectionException =
 				assertThrows(ReflectionException.class, () -> Constructors.IgnoreAccess.newInstance(cls));
 		InvocationTargetException invocationTargetException = JavaObjects.cast(reflectionException.getCause());
 		return JavaObjects.cast(invocationTargetException.getCause());
+	}
+
+	static Properties loadProperties(final String filePath) {
+		try (FileInputStream input = new FileInputStream(filePath)) {
+			Properties properties = new Properties();
+			properties.load(input);
+			return properties;
+		} catch (Exception e) {
+			throw new RuntimeException("Failed to load project properties", e);
+		}
 	}
 
 }